+++ vortex-layout/src/lib.rs
@@ -8,6 +8,7 @@ pub use encoding::*;
 pub use flatbuffers::*;
 pub use layout::*;
 pub use reader::*;
+pub use reader_context::*;
 pub use strategy::*;
 use vortex_session::registry::Context;
 pub use vtable::*;
@@ -18,6 +19,7 @@ mod encoding;
 mod flatbuffers;
 mod layout;
 mod reader;
+mod reader_context;
 pub mod scan;
 pub mod segments;
 pub mod sequence;
